[fxfs] Align file readahead to 64KiB-128KiB boundaries.
Folds the logic up for fxblob and fxfile to both use chunked aligned
readahead. For fxblob which may have non-power of two read sizes it
picks a number between 64KiB and 128KiB that is divisible or if it
exceeds 128KiB, it doesn't do any additional readahead.

Change-Id: Ic32bcada3bc60a2bb701b5e82ef1b49ef41d654e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/947100
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>