[audio][drivers] Remove deprecated external_delay
Previously, an audio driver's 'external_delay' was moved
from the static RingBufferProperties to a dynamic DelayInfo,
returned by the `WatchDelayInfo` hanging-get. Although the
external_delay value in RingBufferProperties was marked as
deprecated, a number of drivers still returned it.

With this CL, no in-tree audio drivers set a value for
RingBufferProperties.external_delay. This is a prerequisite
for actually removing the field from the FIDL table at some
point in the future.

Bug: fxb/116898

Change-Id: Ib350a8bd74e26acb45e75dae0df09d56ff601d29
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/947368
Reviewed-by: Andres Oportus <andresoportus@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>