[component_manager] Factor out shared availability transformation
This is a useful primitive that appears both in production and in
playgrounds. In doing so I also added tests since there's a deficiency
in unit tests in some capability types.

This answers the question of "should we check relative paths" in
new_forwarding_router in the negative.

Change-Id: I0efe98c19036aabb397bdb6ab220d6548f1c9d02
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/947546
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>