[assembly] Add initial Bluetooth platform config
Define an initial configuration for Bluetooth on the system. This config
will be evolved (currently a TODO) to define a comprehensive set of
configuration options for packages in the Bluetooth realm.

Define AIBs for bt-snoop which can either be started eagerly or lazily.
A follow up change will update the product definitions to explicitly
include the snoop AIB if needed.

Define a stub subsystem configuration for the Bluetooth realm. This will
be populated after the product definitions have been updated.

Bug: b/310964410
Bug: b/292109810

Test: fx build -- verify assembly is happy with newly added definition

Change-Id: I82b7072c4a33f358707330b842b59e5b07f17ee1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/946732
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>