[superproject] Roll llvm-project/libc [libc] Add base for target config within cmake (#72318)
Currently the only way to add or remove entrypoints is to modify the
entrypoints.txt file for the current target. This isn't ideal since
a user would have to carry a diff for this file when updating their
checkout. This patch adds a basic mechanism to allow the user to remove
entrypoints without modifying the repository.

GitOrigin-RevId: feaa182b47aa615b354dc91c99db3d73c5bd08c0
Original-Revision: 33a6e7dcaa1b61032d344142511e251be5167963
Roller-URL: https://ci.chromium.org/b/8764142406677684753
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Ie950e82ec6ce163c401f64de9f47542e84839268
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/948577