[a11y] Backdate fuchsia:accessibility::semantics::EnabledState API to v15
This change is using the "escape hatch" described in
go/Smart-Components-Stable-APIs-Escape-Hatch. It is being used because
the use case for b/261482081 falls under use case 2 of "Targeting
In-Development APIs" in go/sd-petals-targeting-stable-apis.

This is a newly added FIDL API that was added in fxr/941818, with no
existing usages.

Bug: b/261482081
Bug: b/311150772
Test: n/a
Change-Id: Ic70064770c26be049d612e97e42ad235c5aa9f26
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/946310
API-Review: Hunter Freyer <hjfreyer@google.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Commit-Queue: Jonny Wang <jonnywang@google.com>