[ffx] Add ssh ConnectTimeout option
Sometimes it appears an emulator can listen on a port, and then
never forward the port to the guest O/S. At least that might
explain why ssh connections to the emulator sometimes just
stall when we try to establish a host-pipe immediately after
the emulator starts. By adding a timeout, we have a chance
to re-try the ssh connection.

This may not work, because the ffx emu's `is_active()`` function checks
for a host-pipe error, not waiting for a retry. But we might at least
get better diagnostics this way, even if it doesn't work.

The timeout has to be chosen somewhat judiciously -- too short and we
_always_ hit the above error.

Bug: b/301312339
Change-Id: Iaa8ff482051a20b48d7ad41649b66f3d6f0096be
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/948621
Commit-Queue: Steven Grady <slgrady@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>