[windowed-stat] Update CombinedWindowedStats implementation
CombinedWindowedStats is now time-aware and automatically slides the
windows if the current time does not fall into existing windows. This
automated operation occurs when the client logs into the
CombinedWindowedStats or dumps it into Inspect.

In this implementation, the bounds of each window are no longer at the
timestamps when windows are slidden. Instead, the bounds of the windows
are always fixed (or prealigned) at the timestamps that divide the
granularity. For example, for the `1m` bucket, the bounds are at the
0th, 60th, 120th second mark, and so on.

CombinedWindowedStats is now renamed to TimeSeries.

Bug: 135511
Test:
1. Verified timestamp is logged
2. Updated unit tests

Change-Id: I4644bbdb8c9c0cc067218e7d25bedcad4f3f3915
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/934810
Reviewed-by: Kevin Sakuma <sakuma@google.com>
Reviewed-by: Devon H. O'Dell <dhobsd@google.com>
Reviewed-by: Antonio Linhart <antoniolinhart@google.com>
Commit-Queue: Kiet Tran <kiettran@google.com>