[fxtest] Support exact match with all modes
Previous to this CL, EXACT was a mode similar to HOST or DEVICE, but now
it is a separate argument. This allows users to express things like
"exact match for host tests," which previously was not possible.

Fixed: b/312271472
Change-Id: Ief2c33fa72773770d2d765ebd055fac96fc205a5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/950079
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Christopher Johnson <crjohns@google.com>