[fxtest] Implement --updateifinbase
This flag was present but did not work properly due to some issues with
the update-if-in-base script. The new implementation properly parses the
base package lists and if necessary will build the update package and
perform an OTA.

During local runs the OTA will fail, so the result is printing a warning
in that case.

Fixed: b/295340410
Change-Id: I31587b74aeb32754ec72a88f1e1cf3f3ce03c1d6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/950204
Fuchsia-Auto-Submit: Christopher Johnson <crjohns@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>