[netcfg] Store provisioning in InterfaceState
Use a fallback provisioning rule to start storing the provisioning
choice specified by the configuration. For purposes of maintaining
current behaviors, the delegating choice for the fallback is dependent
on the value of the install only flag. Once the configuration files are
updated to use the provisioning policy, the fallback provisioning policy
will always specify Local provisioning.

This CL makes no change to the specified install_only flag. That
will be handled in a later CL.

Bug: 135087
Bug: 135160
Bug: b/309010919

Change-Id: Icaf4b75265314063b7ff1851aa5890288385c3f5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/947827
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>