[fxfs] Do not trim read sizes without supplying zeros.
Moves the size correction down after supplying zeroes for requested
regions outside the end of the file. The kernel shouldn't be asking for
those pages anyways, but there is a race with reads and truncation that
can result if this case isn't handled which will result in the pager not
fulfulling the entire request from the kernel and letting it hang.

Change-Id: I5fdd274c930bf623932d31f90e5607c35a0a9b8a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/950643
Fuchsia-Auto-Submit: Martin Lindsay <mlindsay@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>