[netcfg] Add stable name test for non-retry names
In netcfg interface name generation, if only a static name is specified
in the configuration naming policy, there is no capacity to re-try to
generate a new name to produce a different result.

Test that when two devices are added, and will attempt to use the same
name, that netcfg continues to work. For now, we can assert the name on
this interface, but in netcfg's upcoming change to remove temporary
names the test will update to ensure that the second interface either
causes an uninstallation of the first interface or denying the
installation of the second interface.

Linked to change made in fxr/947612

Bug: 135087
Bug: 56559
Bug: b/309010918

Change-Id: I349b5964d4973057a422f65173348a02ffe42aa0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/948613
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>