[netstack3] Access other BufferTransportIpContext
Provide `DualStackDatagramBoundStateContext` with a method that
grants access to a sync_ctx that implements `BufferTransportIpContext`
for the other stack. This will allow IPv6 sockets to send IPv4 packets.

Doing so required implementing `BufferTransportIpContext` for an
uninstantiable type. This type, along with the `UinstantiableCtxMarker`
should be removed once the `BufferFoo` traits are folded into the `Foo`
traits throughout the stack (see https://fxbug.dev/135142).

Bug: 132465
Change-Id: I4c5bd8cffa8468d694e22b5b66e686f80b746070
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/943752
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>