[amlogic-display] Merge the amlogic-hdmitx library into the amlogic-display driver.
The amlogic-hdmitx library is used for the soft transition from
the aml-hdmi driver to the amlogic-display driver, and it's
expected to have the amlogic-display driver as its only client.

Now that the transition is done, we can merge the HdmiTransmitter
class to the amlogic-display driver.

Test: amlogic-hdmitx-test moved to amlogic-display-tests
Test: no-op.
Bug: 134691
Change-Id: I9142b9b578d2a9925e45bf8fe3f27dc6e95add55
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/945639
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Victor Costan <costan@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>
Fuchsia-Auto-Submit: Yilong Li <liyl@google.com>