[fidlc] Unify name collision checking
This CL unifies all name collision checking in the AvailabilityStep.
Previously, fidlc only checked the top-level declarations there, and
delayed member name collision checking to the CompileStep.

This is in preparation for implementing RFC-0231: FIDL versioning
replacement syntax. It will be easier to implement with this name logic
in one place because the rules for the "removed" and "replaced"
arguments are defined in terms of names, and it treats all elements
(i.e. both top-level declarations and members) in the same way.

This also results in better error messages for member name collisions,
because they are now reported before decomposition. For example:

    1 | @available(added=1)
    2 | library example;
    3 | type Foo = struct {
    4 |     @available(added=2) a uint8;
    5 |     @available(added=3) a uint8;
    6 |     @available(added=4) b uint8;
    7 |     @available(added=5) c uint8;
    8 | };

Before, there were 3 duplicate error messages:

    example.fidl:5:25: error: duplicate struct member named 'a'; previous was at example.fidl:4:25
    example.fidl:5:25: error: duplicate struct member named 'a'; previous was at example.fidl:4:25
    example.fidl:5:25: error: duplicate struct member named 'a'; previous was at example.fidl:4:25

Now, there is only 1 error message:

    example.fidl:5:25: error: struct member 'a' has the same name as the struct member declared at example.fidl:4:25; both are available from version 3 onward of platform 'example'

Note that there is still one place in the CompileStep where we have to
check for duplicate member names again: when compiling a protocol, we
must ensure its method names do not collide with composed protocols.
This can't be done in the AvailabilityStep because at that point we
don't know what `P` refers to in `compose P;`.

Test: fx test fidlc-test
Bug: 123989, 135676
Change-Id: Id6ab52a0ed16e7ecd4449cd203e0c2d38bb0a247
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/950274
Fuchsia-Auto-Submit: Mitchell Kember <mkember@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Mitchell Kember <mkember@google.com>